Make manual bindings an optional feature
