+++ notify/Cargo.toml
@@ -12,7 +12,7 @@ keywords = ["events", "filesystem", "notify", "watch"]
 categories = ["filesystem"]
 authors = [
   "Félix Saparelli <me@passcod.name>",
-  "Daniel Faust <hessijames@gmail.com>"
+  "Daniel Faust <hessijames@gmail.com>",
 ]
 
 edition = "2021"
@@ -34,6 +34,18 @@ fsevent-sys = { version = "4", optional = true }
 kqueue = { version = "1.0", optional = true }
 mio = { version = "0.8", features = ["os-ext"], optional = true }
 
+[target.'cfg(windows)'.dependencies.windows-sys]
+version = "0.45.0"
+optional = true
+features = [
+  "Win32_Foundation",
+  "Win32_Security",
+  "Win32_Storage_FileSystem",
+  "Win32_System_IO",
+  "Win32_System_Threading",
+  "Win32_System_WindowsProgramming",
+]
+
 [target.'cfg(any(target_os="freebsd", target_os="openbsd", target_os = "netbsd", target_os = "dragonflybsd"))'.dependencies]
 kqueue = "^1.0.4" # fix for #344
 mio = { version = "0.8", features = ["os-ext"] }
@@ -44,8 +56,9 @@ tempfile = "3.2.0"
 nix = "0.23.1"
 
 [features]
-default = ["macos_fsevent","crossbeam-channel"]
+default = ["macos_fsevent", "crossbeam-channel", "windows-sys"]
 timing_tests = []
 manual_tests = []
 macos_kqueue = ["kqueue", "mio"]
 macos_fsevent = ["fsevent-sys"]
+slim_windows = []
